starting to fix the examples
