+++ cabal.project
@@ -57,21 +57,3 @@ source-repository-package
     type: git
     location: https://github.com/noinia/approximate-equality
     tag: 02fc43ca97dd66940dc3d4ac61ecda7888ec252e
-
--- source-repository-package
---     type: git
---     location: https://github.com/ghcjs/jsaddle
---     subdir: jsaddle
-
--- source-repository-package
---     type: git
---     location: https://github.com/noinia/jsaddle
---     subdir: jsaddle
---     tag: e034fd40cd07c23ad85f3b059e8e3d8b698e322a
-
--- source-repository-package
---     type: git
---     location: https://github.com/noinia/miso
-
--- package miso
---     flags: +jsaddle
