+++ hgeometry-svg/hgeometry-svg.cabal
@@ -58,9 +58,6 @@ common all-setup
       , vinyl                     >= 0.14          && < 1
       , singletons                >= 3             && < 4
       , miso                      >= 1.9           && < 2
-      -- , lucid                     >= 2             && < 3
-      -- , lucid-svg                 >= 0.7           && < 1
-      -- , jsaddle                   >= 0.9.9         && < 0.10
 
 
     default-language: GHC2024
